Fix coverage bugs

+++ run-tests.sh
@@ -21,8 +21,8 @@ set -ex
 dist_mode=$1
 
 python3 -m pip freeze | grep -qi /brainiak \
-        || [ $dist_mode = "--sdist-mode" ] \
-        || [ $dist_mode = "--bdist-mode" ] \
+        || [ ${dist_mode:-default} = "--sdist-mode" ] \
+        || [ ${dist_mode:-default} = "--bdist-mode" ] \
         || {
     echo "You must install brainiak in editable mode"`
         `" before calling "$(basename "$0")
@@ -37,30 +37,36 @@ if [ -z $mpi_command ]; then
 elif [ ! -z $SLURM_NODELIST ]; then
     mpi_command=srun
 fi
-coverage=$(which coverage)
 
-$mpi_command -n 2 $coverage run -m pytest
-
-$coverage combine
-
-# Travis error workaround
-coverage_report=$(mktemp -u coverage_report_XXXXX) || {
-    echo "mktemp -u error" >&2;
-    exit 1;
-}
-
-set +e
-$coverage report > $coverage_report
-report_exit_code=$?
-
-$coverage html
-$coverage xml
-
-cat $coverage_report
-rm $coverage_report
-
-if [ $report_exit_code = 2 ]
+if [ ${dist_mode:-default} = "--bdist-mode" ]
 then
-    echo "ERROR: Coverage too low."
+   $mpi_command -n 2 $(which pytest)
+
+# Only run coverage for editable or source distribution
+else
+   $mpi_command -n 2 $(which coverage) run -m pytest
+   coverage combine
+
+   # Travis error workaround
+   coverage_report=$(mktemp -u coverage_report_XXXXX) || {
+      echo "mktemp -u error" >&2;
+      exit 1;
+   }
+
+   set +e
+   coverage report > $coverage_report
+   report_exit_code=$?
+
+   coverage html
+   coverage xml
+
+   cat $coverage_report
+   rm $coverage_report
+
+   if [ $report_exit_code = 2 ]
+   then
+      echo "ERROR: Coverage too low."
+   fi
+   exit $report_exit_code
 fi
-exit $report_exit_code
+
